feat: Move safe_async_cache.py to async_components/cache/
